Bug fix There was a problem introduced in the previous commit.

+++ docs/penrose.js
@@ -327,7 +327,7 @@ export class VertexGroup {
             const dartOut = [];
             this.vertices.forEach((vertex) => {
                 if (vertex.type == "dart") {
-                    if (vertex.to.short) {
+                    if (vertex.to.long) {
                         dartIn.push(vertex);
                     }
                     else {
